Move LogRoot and MapRoot to types package
