+++ crypto/data_formats_test.go
@@ -16,7 +16,6 @@ package crypto
 
 import (
 	"bytes"
-	"reflect"
 	"testing"
 
 	"github.com/google/trillian"
@@ -24,81 +23,6 @@ import (
 )
 
 var dh = testonly.MustHexDecode
-
-func TestLogRoot(t *testing.T) {
-	for _, logRoot := range []*LogRootV1{
-		{
-			RootHash: []byte("foo"),
-			Metadata: []byte{},
-		},
-	} {
-		b, err := SerializeLogRoot(logRoot)
-		if err != nil {
-			t.Errorf("SerializeLogRoot(%v): %v", logRoot, err)
-		}
-		got, err := ParseLogRoot(b)
-		if err != nil {
-			t.Errorf("ParseLogRoot(): %v", err)
-		}
-		if !reflect.DeepEqual(got, logRoot) {
-			t.Errorf("serialize/parse round trip failed. got %#v, want %#v", got, logRoot)
-		}
-	}
-}
-
-func TestParseLogRoot(t *testing.T) {
-	for _, tc := range []struct {
-		logRoot []byte
-		wantErr bool
-	}{
-		{
-			logRoot: func() []byte {
-				b, _ := SerializeLogRoot(&LogRootV1{})
-				return b
-			}(),
-		},
-		{
-			logRoot: func() []byte {
-				b, _ := SerializeLogRoot(&LogRootV1{})
-				b[0] = 1 // Corrupt the version tag.
-				return b
-			}(),
-			wantErr: true,
-		},
-		{
-			logRoot: []byte("foo"),
-			wantErr: true,
-		},
-	} {
-		_, err := ParseLogRoot(tc.logRoot)
-		if got, want := err != nil, tc.wantErr; got != want {
-			t.Errorf("ParseLogRoot(): %v, wantErr: %v", err, want)
-		}
-	}
-}
-
-func TestMapRoot(t *testing.T) {
-	for _, tc := range []struct {
-		mapRoot *MapRootV1
-	}{
-		{mapRoot: &MapRootV1{
-			RootHash: []byte("foo"),
-			Metadata: []byte{},
-		}},
-	} {
-		b, err := SerializeMapRoot(tc.mapRoot)
-		if err != nil {
-			t.Errorf("SerializeMapRoot(%v): %v", tc.mapRoot, err)
-		}
-		got, err := ParseMapRoot(b)
-		if err != nil {
-			t.Errorf("ParseMapRoot(): %v", err)
-		}
-		if !reflect.DeepEqual(got, tc.mapRoot) {
-			t.Errorf("serialize/parse round trip failed. got %#v, want %#v", got, tc.mapRoot)
-		}
-	}
-}
 
 // It's important that signatures don't change.
 func TestHashLogRootKnownValue(t *testing.T) {
